Add types, make strategy names consumable
This commit adds additional typing without affecting the structure
of the strategy definitions. While it may be necessary to refactor
the definitions to have a dedicated `args` property in the future,
this commit simply adds some more developer sugar to make things
easier to work with. This will still not perform validation of
arguments.